+++ app/examples/basic-configuration/BasicConfigurationExample.tsx
@@ -1,33 +1,10 @@
-'use client';
-
-import { useContextMenu } from '__PACKAGE__';
-import { Picture } from '~/components/Picture';
-import { copyImageToClipboard, downloadImage, unsplashImages } from '~/lib/images';
+import { BasicConfigurationExampleContent } from './BasicConfigurationExampleContent';
+import { ProviderPropsExample } from './ProviderPropsExample';
 
 export function BasicConfigurationExample() {
-  // example-start
-  const showContextMenu = useContextMenu();
-  // example-skip
-  const image = unsplashImages[2];
-  const { src } = image.file;
-  // example-resume
   return (
-    <Picture
-      image={image}
-      onContextMenu={showContextMenu(
-        [
-          {
-            key: 'copy',
-            onClick: () => copyImageToClipboard(src),
-          },
-          {
-            key: 'download',
-            onClick: () => downloadImage(src),
-          },
-        ],
-        { zIndex: 1000, shadow: 'md', borderRadius: 'md' }
-      )}
-    />
+    <ProviderPropsExample>
+      <BasicConfigurationExampleContent />
+    </ProviderPropsExample>
   );
-  // example-end
 }