+++ app/examples/submenus/SubmenuDelayExample.tsx
@@ -1,0 +1,10 @@
+import { SubmenuDelayProviderPropExample } from './SubmenuDelayProviderPropExample';
+import { SubmenuExample } from './SubmenuExamples';
+
+export function SubmenuDelayExample() {
+  return (
+    <SubmenuDelayProviderPropExample>
+      <SubmenuExample />
+    </SubmenuDelayProviderPropExample>
+  );
+}