+++ app/getting-started/page.tsx
@@ -1,5 +1,5 @@
 import { Code } from '@mantine/core';
-import { MANTINE_LINK, PACKAGE_NAME, PRODUCT_NAME } from '~/app/config';
+import { MANTINE_LINK, PRODUCT_NAME } from '~/app/config';
 import { CodeBlock } from '~/components/CodeBlock';
 import { ExternalLink } from '~/components/ExternalLink';
 import { InternalLink } from '~/components/InternalLink';
@@ -31,18 +31,18 @@ export default async function GettingStartedPage() {
         {PRODUCT_NAME} depends on <Code>@mantine/core</Code>, <Code>@mantine/hooks</Code> and <Code>clsx</Code>.
       </Txt>
       <Txt>
-        Create a new <ExternalLink to={`${MANTINE_LINK}pages/getting-started/`}>application with Mantine</ExternalLink>{' '}
-        and install the package:
+        Create a new <ExternalLink to={`${MANTINE_LINK}/getting-started/`}>application with Mantine</ExternalLink> and
+        install the package:
       </Txt>
       <CodeBlock
         tabs={[
           {
             language: 'shell',
             fileName: 'yarn',
-            code: `yarn add ${PACKAGE_NAME}`,
+            code: `yarn add ${process.env.PACKAGE_NAME} clsx`,
           },
-          { language: 'shell', fileName: 'npm', code: `npm i ${PACKAGE_NAME}` },
-          { language: 'shell', fileName: 'pnpm', code: `pnpm i ${PACKAGE_NAME}` },
+          { language: 'shell', fileName: 'npm', code: `npm i ${process.env.PACKAGE_NAME} clsx` },
+          { language: 'shell', fileName: 'pnpm', code: `pnpm i ${process.env.PACKAGE_NAME} clsx` },
         ]}
       />
       <Txt>