+++ app/layout.tsx
@@ -3,14 +3,14 @@ import { Notifications } from '@mantine/notifications';
 import { ContextMenuProvider } from '__PACKAGE__';
 import { Metadata } from 'next';
 import { AppWrapper } from '~/components/AppWrapper';
-import { WEBSITE_LINK } from './config';
+import { PRODUCT_NAME, WEBSITE_LINK } from './config';
 import './layout.css';
 import classes from './layout.module.css';
 
 export const metadata: Metadata = {
-  metadataBase: process.env.GITHUB_PAGES === 'TRUE' ? new URL('https://icflorescu.github.io') : undefined,
+  metadataBase: new URL(process.env.GITHUB_PAGES === 'TRUE' ? 'https://icflorescu.github.io' : 'http://localhost:3000'),
   manifest: `${process.env.GITHUB_PAGES === 'TRUE' ? WEBSITE_LINK : ''}/manifest.webmanifest`,
-  title: 'Mantine ContextMenu: enhance your users’ experience',
+  title: `${PRODUCT_NAME}: enhance your users’ experience`,
   description:
     'Design your Mantine applications for productivity and meet your users’ expectations by enhancing your UIs with desktop-grade context menus',
 };