+++ lib/useNpmDownloads.ts
@@ -7,7 +7,7 @@ const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
 export function useNpmDownloads() {
   const { data } = useSWR<{ downloads: number }>(
-    'https://api.npmjs.org/downloads/point/last-month/mantine-contextmenu',
+    `https://api.npmjs.org/downloads/point/last-month/${process.env.PACKAGE_NAME}`,
     fetcher,
     {
       refreshInterval: DOWNLOADS_REFRESH_INTERVAL,