+++ package/ContextMenuItem.tsx
@@ -1,9 +1,11 @@
 'use client';
 
 import { Box, UnstyledButton, parseThemeColor, rgba, useMantineTheme } from '@mantine/core';
+import { useTimeout } from '@mantine/hooks';
 import clsx from 'clsx';
-import { useRef, useState, type MouseEventHandler } from 'react';
+import { useContext, useRef, useState, type MouseEventHandler } from 'react';
 import { ContextMenu } from './ContextMenu';
+import { MenuSettingsContext } from './ContextMenuProvider';
 import type { ContextMenuContent, ContextMenuItemOptions, WithRequiredProperty } from './types';
 
 export function ContextMenuItem({
@@ -18,6 +20,8 @@ export function ContextMenuItem({
   items,
 }: WithRequiredProperty<Omit<ContextMenuItemOptions, 'key'>, 'title'> & { onHide: () => void }) {
   const ref = useRef<HTMLButtonElement>(null);
+  const { submenuDelay } = useContext(MenuSettingsContext);
+
   const [submenuPosition, setSubmenuPosition] = useState<{ x: number; y: number } | null>(null);
 
   const handleClick: MouseEventHandler<HTMLButtonElement> | undefined = onClick
@@ -27,13 +31,23 @@ export function ContextMenuItem({
       }
     : undefined;
 
-  const showSubmenu = () => {
+  const { start: startShowingSubmenu, clear: stopShowingSubmenu } = useTimeout(() => {
     const { top: y, right: x } = ref.current!.getBoundingClientRect();
     setSubmenuPosition({ x, y });
+  }, submenuDelay);
+
+  const { start: startHidingSubmenu, clear: stopHidingSubmenu } = useTimeout(() => {
+    setSubmenuPosition(null);
+  }, submenuDelay);
+
+  const showSubmenu = () => {
+    stopHidingSubmenu();
+    startShowingSubmenu();
   };
 
   const hideSubmenu = () => {
-    setSubmenuPosition(null);
+    stopShowingSubmenu();
+    startHidingSubmenu();
   };
 
   const hasItemsAndIsNotDisabled = items && !disabled;
@@ -41,12 +55,11 @@ export function ContextMenuItem({
   const theme = useMantineTheme();
   const { colors } = theme;
   const parsedColor = color ? parseThemeColor({ color, theme }).value : undefined;
-  if (color === 'red') console.log(parsedColor);
 
   return (
     <div
-      onMouseOver={hasItemsAndIsNotDisabled ? showSubmenu : undefined}
-      onMouseOut={hasItemsAndIsNotDisabled ? hideSubmenu : undefined}
+      onMouseEnter={hasItemsAndIsNotDisabled ? showSubmenu : undefined}
+      onMouseLeave={hasItemsAndIsNotDisabled ? hideSubmenu : undefined}
     >
       <UnstyledButton
         ref={ref}